add integration tests that use the `--copy-unsafe` argument
Each `for_each_xxx` macro can now take parameters in addition to
the `test_one` function to run on each test case. These additional
parameters are passed on to the `test_one` when that method is
called.

Most `test_one` functions now take an `extra_args` parameter, which
is used so that select test cases can use the `--copy-unsafe`
argument without all test cases using that argument.

Adds images which have private unsafe-to-copy chunks for the test
cases to use.